Update code for UI kitchen

+++ Menu-Service/src/main/resources/schema.sql
@@ -1,5 +1,5 @@
 CREATE TABLE menus (
-                       menu_id INT AUTO_INCREMENT PRIMARY KEY,
+                       menu_id BIGINT AUTO_INCREMENT PRIMARY KEY,
                        name VARCHAR(255) NOT NULL,
                        description TEXT,
                        imageUrl VARCHAR(100),
